#  Performance update for 1.0 