… Demo/Measures/AdultOutpatientEncountersFHIR4-2.2.000.cs
… Demo/Measures/AdvancedIllnessandFrailtyExclusionECQMFHIR4-5.17.000.cs
… Demo/Measures/BCSEHEDISMY2022-1.0.0.cs
… Demo/Measures/BreastCancerScreeningsFHIR-0.0.009.cs
… Demo/Measures/CervicalCancerScreeningFHIR-0.0.005.cs
… Demo/Measures/ColorectalCancerScreeningsFHIR-0.0.003.cs
… Demo/Measures/CumulativeMedicationDurationFHIR4-1.0.000.cs
… Demo/Measures/DRCommunicationWithPhysicianManagingDiabetesFHIR-0.0.004.cs
… Demo/Measures/DevDays-2023.0.0.cs
… Demo/Measures/DiabetesHemoglobinA1cHbA1cPoorControl9FHIR-0.0.015.cs
… Demo/Measures/DischargedonAntithromboticTherapyFHIR-0.0.010.cs
… Demo/Measures/Exam125FHIR-0.0.009.cs
… Demo/Measures/Exam130FHIR-0.0.003.cs
… Demo/Measures/FHIR347-0.1.021.cs
… Demo/Measures/FHIRHelpers-4.0.001.cs
… Demo/Measures/HospiceFHIR4-2.3.000.cs
… Demo/Measures/HospitalHarmHyperglycemiainHospitalizedPatientsFHIR-0.0.006.cs
… Demo/Measures/HospitalHarmSevereHypoglycemiaFHIR-0.0.012.cs
… Demo/Measures/HybridHWMFHIR-0.102.005.cs
… Demo/Measures/HybridHWRFHIR-1.3.005.cs
… Demo/Measures/MATGlobalCommonFunctionsFHIR4-6.1.000.cs
… Demo/Measures/NCQAAdvancedIllnessandFrailty-1.0.0.cs
… Demo/Measures/NCQAClaims-1.0.0.cs
… Demo/Measures/NCQAEncounter-1.0.0.cs
… Demo/Measures/NCQAFHIRBase-1.0.0.cs
… Demo/Measures/NCQAHealthPlanEnrollment-1.0.0.cs
… Demo/Measures/NCQAHospice-1.0.0.cs
… Demo/Measures/NCQAPalliativeCare-1.0.0.cs
… Demo/Measures/NCQAStatus-1.0.0.cs
… Demo/Measures/NCQATerminology-1.0.0.cs