- restore the lazy loading and constructors but only for measures with context defined
- add lazy null handling
- addresses #588

Current 1.0

ADD-E: Duration: 3h 32m 35s
CRE: Duration: 2h 46m 1s
PND-E: Duration: 10h 48m 39s
SPD: Duration: 3h 10m 21s

Time to run 80 HEDIS measures with multiple configurations using current develop 1.0 branch
![image](https://github.com/user-attachments/assets/0acdef5d-36f3-4a34-a0b0-529a9d57e517)

With these updates, same build pipeline

ADD-E: Duration: 38m 0s
CRE: Duration: 57m 4s
PND-E: Duration: 53m 39s
SPD: Duration: 1h 36m 22s

![image](https://github.com/user-attachments/assets/9a7a566b-a161-43a1-b37a-d7006c0ac49a)