@DamienMosier 2024-10-08
I could use some feedback on why the CQL Test exclusion is failing. The CQL and demo project build and the unit tests pass in the Demo project.