@baseTwo 2024-10-17
It looks like the generated c# code in the Demo project reverted to the old way? Could these be regenerated as part of this PR to demonstrate how they ought to look like now?