@baseTwo 2024-10-17
> @baseTwo - is this something we need in your equivalent 2.0 port as well?

It looks like 1.0 changed back to the old way where libraries contain state for the context and the Lazy<>'s ? 

@DamienMosier , which measures do not have context defined, I'd like to compare the two versions, but GitHub collapses too many files for me to go through each one