@DamienMosier 2024-10-17
@ewoutkramer this is something you would definitely want in the 2.0 branch for the performance gains.

@baseTwo The way the logic work is if the file or one of the includes has a context then it creates it the old way with a constructor and utilizes Lazy<T>. If the file does not, aka a library, then it's a singleton. 

For our use case only our actual measures have context Patient and all of our libraries are either functions or fluent functions with the resources passed in. In your Demo project, context is added to almost all files which is why it looks like the old way.

The main performance issue with the original 1.0 branch was with the constructors. If you have LibraryA that has 5 includes and each of those subsequent libraries each have their own set of includes. You could end up instantiating the same library multiple times. Having it scoped allowed us to use Lazy<T> so we gained the benefit of some caching but at the expense of calling the constructors a lot. Switching to singleton removed that constructor lift but removed the ability to use Lazy<T>. I thought about it and realized we only need Lazy<T> where there is context and all subsequent libraries did not need it and we could have the best of both worlds.

Here are some partial examples of measure AAB from DQIC if you want to see the CQL.

```
[System.CodeDom.Compiler.GeneratedCode(".NET Code Generation", "1.0.0.0")]
[CqlLibrary("AAB_Concepts", "2024.1.0")]
public class AAB_Concepts_2024_1_0
{

    public static AAB_Concepts_2024_1_0 Instance { get; }  = new();

    #region Dependencies
    #endregion

    [CqlDeclaration("Acute Bronchitis")]
    [CqlValueSet("https://www.ncqa.org/fhir/valueset/2.16.840.1.113883.3.464.1004.1016")]
	public CqlValueSet Acute_Bronchitis(CqlContext context) => 
		new CqlValueSet("https://www.ncqa.org/fhir/valueset/2.16.840.1.113883.3.464.1004.1016", null);

    [CqlDeclaration("Competing Diagnosis")]
    [CqlValueSet("https://www.ncqa.org/fhir/valueset/2.16.840.1.113883.3.464.1004.1067")]
	public CqlValueSet Competing_Diagnosis(CqlContext context) => 
		new CqlValueSet("https://www.ncqa.org/fhir/valueset/2.16.840.1.113883.3.464.1004.1067", null);

```
and the actual file with context is reverted to the old way, the same way you see files in the DQIC repository.

```
[System.CodeDom.Compiler.GeneratedCode(".NET Code Generation", "1.0.0.0")]
[CqlLibrary("AAB_Reporting", "2024.1.0")]
public class AAB_Reporting_2024_1_0
{
    internal CqlContext context;

    #region Cached values

    internal Lazy<Tuples.Tuple_BOFQdLfCiMegGCeQZaSIJRaC> __Measure;
    internal Lazy<Patient> __Patient;
    internal Lazy<IEnumerable<ExplanationOfBenefit>> __Eobs;
    internal Lazy<IEnumerable<ExplanationOfBenefit>> __Admin_Eobs;
    
    #endregion
    public AAB_Reporting_2024_1_0(CqlContext context)
    {
        this.context = context ?? throw new ArgumentNullException("context");
        
        __Measure = new Lazy<Tuples.Tuple_BOFQdLfCiMegGCeQZaSIJRaC>(this.Measure_Value(context));
        __Patient = new Lazy<Patient>(this.Patient_Value(context));
        __Eobs = new Lazy<IEnumerable<ExplanationOfBenefit>>(this.Eobs_Value(context));
        __Admin_Eobs = new Lazy<IEnumerable<ExplanationOfBenefit>>(this.Admin_Eobs_Value(context));
```