@DamienMosier 2024-10-17
For reference, I generated this for Evan doing testing. The number are based off our build machine times. The "1.0 DCS" is the original 1.0 prior to singleton. The "1.0 Develop" is current develop with only singleton's and the final are these changes that are pending.

![image](https://github.com/user-attachments/assets/32f2907b-e262-4268-b035-e45ffd273a8c)